+++ src/cks/constraints/__init__.py
@@ -1,5 +1,5 @@
 from .registry import registry
-from .builtin import BUILTIN_CONSTRAINTS, OPTIONAL_CONSTRAINTS
+from .builtin import BUILTIN_CONSTRAINTS, OPTIONAL_CONSTRAINTS, OPTIONAL_CONSTRAINTS_BY_NAME
 
 for constraint in BUILTIN_CONSTRAINTS:
     registry.register(constraint)
@@ -13,4 +13,5 @@ __all__ = [
     "registry",
     "BUILTIN_CONSTRAINTS",
     "OPTIONAL_CONSTRAINTS",
+    "OPTIONAL_CONSTRAINTS_BY_NAME",
 ]